% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oft.R
\docType{data}
\name{oft_aq_cyanobacteria}
\alias{oft_aq_cyanobacteria}
\title{Aquatic Cyanobacteria}
\format{
An object of class \code{character} of length 4.
}
\usage{
oft_aq_cyanobacteria
}
\description{
Aquatic Cyanobacteria
}
\examples{
print(oft_aq_cyanobacteria)
}
\seealso{
Other oft_aq: 
\code{\link{oft_aq_diatom_algae}},
\code{\link{oft_aq_fish}},
\code{\link{oft_aq_green_algae}},
\code{\link{oft_aq_invertebrates}},
\code{\link{oft_aq_macrophytes}},
\code{\link{oft_aq_primary_producers}}
}
\concept{oft_aq}
\keyword{datasets}
