% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oft.R
\docType{data}
\name{oft_aq_primary_producers}
\alias{oft_aq_primary_producers}
\title{Aquatic primary producers}
\format{
An object of class \code{character} of length 24.
}
\usage{
oft_aq_primary_producers
}
\description{
Aquatic primary producers
}
\examples{
print(oft_aq_primary_producers)
}
\seealso{
Other oft_aq: 
\code{\link{oft_aq_cyanobacteria}},
\code{\link{oft_aq_diatom_algae}},
\code{\link{oft_aq_fish}},
\code{\link{oft_aq_green_algae}},
\code{\link{oft_aq_invertebrates}},
\code{\link{oft_aq_macrophytes}}
}
\concept{oft_aq}
\keyword{datasets}
