% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oft.R
\docType{data}
\name{oft_cyanobacteria}
\alias{oft_cyanobacteria}
\title{Cyanobacteria}
\format{
An object of class \code{character} of length 4.
}
\usage{
oft_cyanobacteria
}
\description{
Cyanobacteria
}
\examples{
print(oft_cyanobacteria)
}
\keyword{datasets}
