% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oft.R
\docType{data}
\name{oft_diatom_algae}
\alias{oft_diatom_algae}
\title{Diatom species}
\format{
An object of class \code{character} of length 2.
}
\usage{
oft_diatom_algae
}
\description{
Diatom species
}
\examples{
print(oft_diatom_algae)
}
\keyword{datasets}
