% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oft_species_group.R
\name{oft_group_species}
\alias{oft_group_species}
\title{Group species names occurring in reference points}
\usage{
oft_group_species(species)
}
\arguments{
\item{species}{The species name, i.e. the Species column in the referenec
point table}
}
\value{
A species group name
}
\description{
Convenience function to group species into fish, aquatic invertebrates,
aquatic primary producers, birds, mammals, bees, earthworms, arthropods
and plants. The entries 'Unspecified' and 'not reported' are translated
to NA.
}
\examples{
library(dplyr)
oft$reference_points |>
  dplyr::filter(Substance == "Thiacloprid") |>
  mutate(Species_group = oft_group_species(Species)) |>
  select(Substance, Species, Species_group, Endpoint, qualifier, value, unit, Effect)
}
