% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oft_species_group.R
\docType{data}
\name{oft_other}
\alias{oft_other}
\title{Other}
\format{
An object of class \code{character} of length 2.
}
\usage{
oft_other
}
\description{
Other
}
\examples{
print(oft_other)
}
\seealso{
Other oft_species_group: 
\code{\link{oft_aq_cyanobacteria}},
\code{\link{oft_aq_diatom_algae}},
\code{\link{oft_aq_fish}},
\code{\link{oft_aq_green_algae}},
\code{\link{oft_aq_invertebrates}},
\code{\link{oft_aq_macrophytes}},
\code{\link{oft_aq_primary_producers}},
\code{\link{oft_arthropods}},
\code{\link{oft_bees}},
\code{\link{oft_birds}},
\code{\link{oft_earthworms}},
\code{\link{oft_mammals}},
\code{\link{oft_plants}},
\code{\link{oft_soil_macroorganisms}}
}
\concept{oft_species_group}
\keyword{datasets}
