% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oft.R
\docType{data}
\name{oft}
\alias{oft}
\title{Connected tables of data published in OpenFoodTox}
\format{
list A \link{dm} object
}
\source{
Carnesecchi E, Mostrag A, Ciacci A, Roncaglioni A, Tarkhov A, Gibin D,
Sartori L, Benfenati E, Yang C, Dorne JLCM (2023). OpenFoodTox: EFSA's
chemical hazards database (Version 5). Zenodo. \doi{10.5281/zenodo.8120114}

Dorne JLCM, Richardson J, Livaniou A, Carnesecchi E, Ceriani L, Baldin R,
Kovarich S, Pavan M, Saouter E, Biganzoli F, Pasinato L, Zare Jeddi M,
Robinson TP, Kass GEN, Liem AKD, Toropov AA, Toropova AP, Yang C, Tarkhov A,
Georgiadis N, Di Nicola MR, Mostrag A, Verhagen H, Roncaglioni A, Benfenati
E, Bassan A. EFSA's OpenFoodTox: An open source toxicological database on
chemicals in food and feed and its future developments. Environ Int. 2021
Jan;146:106293. \doi{10.1016/j.envint.2020.106293}
}
\usage{
oft
}
\description{
This data object was created to facilitate the access to the list of so-called
'reference points' published as part of the OpenFoodTox datasets (Carnesecchi
et al. 2023). These reference points are linked to substance
characterisations and the outputs of the European Food Safety Authority
(EFSA) in which the data were published. Also, a table of 'reference values'
is included, which is also linked to substance characterisations and EFSA
outputs. Other data tables published as part of OpenFoodTox are not included
in this R package.
}
\details{
See \code{vignette("OpenFoodTox")} for a description of the contents of the individual
tables.

The R code used to create this data object is installed with this package
in the 'dataset_generation' directory.
}
\examples{
library(dm, warn.conflicts = FALSE)
library(dplyr, warn.conflicts = FALSE)

# Show the relational structure of the data tables (only works in online HTML help)
dm_draw(oft, view_type = "all")

# List species used in aquatic tests
oft$reference_points |>
  dplyr::filter(Study == "Ecotox (water compartment)") |>
  select(Species) |>
  unique() |>
  arrange(Species) |>
  print(n = Inf)

# Collect endpoints on green algae
green_algae_endpoints <- oft$reference_points |>
  dplyr::filter(Species \%in\% oft_aq_green_algae) |>
  select(Substance, Species, DurationDays, Endpoint, qualifier, value, unit, Effect)
print(green_algae_endpoints)

# Show aquatic endpoints for spinosad
oft$reference_points |>
  dplyr::filter(Study == "Ecotox (water compartment)") |>
  dplyr::filter(grepl("^Spinos", Substance)) |>
  arrange(Species) |>
  select(Substance, Species, DurationDays, Endpoint, qualifier, value, unit, Effect)

# Check substance characterisation of spinosad (no useful information available)
oft$substance_characterisation |>
  dplyr::filter(grepl("^Spinos", Substance))

# List species used in terrestrial tests
oft$reference_points |>
  dplyr::filter(Study == "Ecotox (soil compartment)") |>
  select(Species) |>
  unique() |>
  arrange(Species) |>
  print(n = Inf)

# Show terrestrial endpoints for spinosad
oft$reference_points |>
  dplyr::filter(Study == "Ecotox (soil compartment)") |>
  dplyr::filter(grepl("^Spinos", Substance)) |>
  arrange(Species) |>
  select(Substance, Species, DurationDays, Endpoint, qualifier, value, unit, Effect)

# Show terrestrial endpoints for terbuthylazine 
oft$reference_points |>
  dplyr::filter(Study == "Ecotox (soil compartment)") |>
  dplyr::filter(grepl("^Terbuthyl", Substance)) |>
  arrange(Species) |>
  select(Substance, Species, DurationDays, Endpoint, qualifier, value, unit, Effect)

# Show terrestrial endpoints for florasulam (an example where terrestrial plant data are available)
oft$reference_points |>
  dplyr::filter(Study == "Ecotox (soil compartment)") |>
  dplyr::filter(grepl("^Floras", Substance)) |>
  arrange(Species) |>
  select(Substance, Species, DurationDays, Endpoint, qualifier, value, unit, Effect)
}
\seealso{
species_group
}
\keyword{datasets}
