% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oft.R
\docType{data}
\name{oft_aq_fish}
\alias{oft_aq_fish}
\title{Aquatic fish}
\format{
An object of class \code{character} of length 21.
}
\usage{
oft_aq_fish
}
\description{
Aquatic fish
}
\examples{
print(oft_aq_fish)
}
\seealso{
Other oft_aq: 
\code{\link{oft_aq_cyanobacteria}},
\code{\link{oft_aq_diatom_algae}},
\code{\link{oft_aq_green_algae}},
\code{\link{oft_aq_invertebrates}},
\code{\link{oft_aq_macrophytes}},
\code{\link{oft_aq_primary_producers}}
}
\concept{oft_aq}
\keyword{datasets}
